Add Contribution Guideline
This pull request adds a new contribution guideline to the project, which provides guidance for contributors on how to contribute especially when the Devstudio are starting. This is can be 

The contribution guideline covers the following topics:

- How to create a new issue
- How to solve an issue
- How to add pull request

If there's something that can be improved or added please let me know, I would love to comply.